requirements.txt

huggingface_hub
ipywidgets
groq
pyspellchecker
streamlit
langchain_core
langchain
langchain-community
sentence-transformers
faiss-cpu
pymongo
kafka-python

